+++ src/App.css
@@ -127,11 +127,12 @@ nav ul li{
 
 /**********************PARTICLES JS ***********************************/
 .particles-canvas{
-  position: absolute;
+  position: absolute ;
   top: 0;
   left: 0;
   right: 0;
   bottom: 0;
+
 
 }
 
@@ -165,7 +166,7 @@ nav ul li{
   display: flex;
   justify-content: center;
   align-items: center;
-  
+
 }
 
 .profile-img{
