+++ src/components/Header.css
@@ -1,0 +1,11 @@
+#header-wraper .particles-canvas{
+    position: absolute ;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    margin-bottom: 2rem;
+ 
+    
+  
+  }
